Update the way we assert for the contact collection details
