+++ test/end-to-end/cypress/specs/DIT/event-spec.js
@@ -1,20 +1,6 @@
 const selectors = require('../../../../selectors')
-const { keys, forEach } = require('lodash')
 
-// will refactor this later, so it's not duplicate from functional support assertion
-const assertKeyValueTable = (dataAutoId, expected) => {
-  forEach(keys(expected), (key, i) => {
-    const rowNumber = i + 1
-    cy.get(selectors.keyValueTable(dataAutoId).keyCell(rowNumber)).should('have.text', key)
-
-    if (expected[key].href) {
-      cy.get(selectors.keyValueTable(dataAutoId).valueCellLink(rowNumber)).should('have.attr', 'href', expected[key].href)
-      cy.get(selectors.keyValueTable(dataAutoId).valueCellLink(rowNumber)).should('have.text', expected[key].name)
-    } else {
-      cy.get(selectors.keyValueTable(dataAutoId).valueCell(rowNumber)).should('have.text', expected[key])
-    }
-  })
-}
+const { assertKeyValueTable } = require('../../support/assertions')
 
 const createEvent = () => {
   cy.get(selectors.eventCreate.eventName).type('Eventful event')
