+++ test/end-to-end/cypress/support/assertions.js
@@ -1,3 +1,7 @@
+const selectors = require('../../../selectors')
+
+const { keys, forEach } = require('lodash')
+
 const assertError = (message) => {
   cy.get('body').should('contain', message)
 }
@@ -7,6 +11,20 @@ const assertCollection = (headerCountSelector, collectionItemsSelector) => {
     cy.get(collectionItemsSelector).should((collectionItems) => {
       expect(headerCount).to.eq(collectionItems.length.toString())
     })
+  })
+}
+
+const assertKeyValueTable = (dataAutoId, expected) => {
+  forEach(keys(expected), (key, i) => {
+    const rowNumber = i + 1
+    cy.get(selectors.keyValueTable(dataAutoId).keyCell(rowNumber)).should('have.text', key)
+
+    if (expected[key].href) {
+      cy.get(selectors.keyValueTable(dataAutoId).valueCellLink(rowNumber)).should('have.attr', 'href', expected[key].href)
+      cy.get(selectors.keyValueTable(dataAutoId).valueCellLink(rowNumber)).should('have.text', expected[key].name)
+    } else {
+      cy.get(selectors.keyValueTable(dataAutoId).valueCell(rowNumber)).should('have.text', expected[key])
+    }
   })
 }
 
@@ -23,4 +41,5 @@ module.exports = {
   assertError,
   assertCollection,
   assertLocalNav,
+  assertKeyValueTable,
 }
